Support evenly distributed pw and choice of hash
Version 2 of the getpw function. This allows to generate
password that are really evenly distributed along the space of
possible passwords. The idea here is that there can be "leading
zeroes" in the first max-min characters. This algorithm will
nearly always generate passwords of max length, so it is optional
and must be switched on by using the flag FL_EVENDIST.

Also, a different hash algorithm can be specified.